refactor(tests): rename discovery helpers.py to conftest.py
Move shared test utilities (FIXTURES_DIR, create_mock_path) from helpers.py
to conftest.py, consistent with the existing convention across the test tree
(5 other conftest.py files).

PR #657 review item #2
